Factored widgets from params_form into gui_main; created a new module for params_form
